Rename some legacy exports
