+++ packages/satcheljs/lib/legacy/LegacyDispatchFunction.ts
