+++ packages/satcheljs/lib/legacy/dispatch.ts
@@ -1,8 +1,8 @@
 import {action as mobxAction} from 'mobx';
 import ActionContext from './ActionContext';
 import ActionFunction from './ActionFunction';
-import DispatchFunction from './DispatchFunction';
-import { dispatchWithMiddleware } from './applyMiddleware';
+import LegacyDispatchFunction from './LegacyDispatchFunction';
+import { dispatchWithMiddleware } from './legacyApplyMiddleware';
 import { getGlobalContext } from '../globalContext';
 
 export default function dispatch(action: ActionFunction, actionType: string, args: IArguments,  actionContext: ActionContext): void {
